+++ ItunesMedia/Library/Protocols/ViewModel/SkeletonCellViewModel.swift
@@ -1,0 +1,11 @@
+//
+//  SkeletonCellViewModel.swift
+//  ItunesMedia
+//
+//  Created by Gustavo Lopez on 12/27/19.
+//  Copyright © 2019 Gustavo Lopez. All rights reserved.
+//
+
+import IGListKit
+
+protocol SkeletonCellViewModel: ListDiffable {}