+++ ItunesMedia/Library/Protocols/ViewModel/TextBasedCellViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  TextBasedCellViewModel.swift
+//  ItunesMedia
+//
+//  Created by Gustavo Lopez on 12/27/19.
+//  Copyright © 2019 Gustavo Lopez. All rights reserved.
+//
+
+import Foundation
+
+protocol TextBasedCellViewModel {
+    var title: String { get }
+    var content: String { get }
+}