+++ ItunesMedia/ViewModels/CellViewModels/TextBasedCellViewModelType.swift
@@ -1,5 +1,5 @@
 //
-//  LoadinCellViewModel.swift
+//  TextBasedCellViewModel.swift
 //  ItunesMedia
 //
 //  Created by Gustavo Lopez on 12/27/19.
@@ -8,33 +8,38 @@
 
 import IGListKit
 
-final class LoadingCellViewModel {
+final class TextBasedCellViewModelType: TextBasedCellViewModel {
     
     // MARK: - Attributes
     
     let identifier: String
+    let title: String
+    let content: String
     
     // MARK: - LifeCycle
     
-    init() {
+    init(title: String, content: String) {
+        self.title = title
+        self.content = content
         identifier = UUID().uuidString
-    }    
+    }
     
 }
 
 // MARK: - ListDiffable
 
-extension LoadingCellViewModel: ListDiffable {
+extension TextBasedCellViewModelType: ListDiffable {
     
     func diffIdentifier() -> NSObjectProtocol {
         return identifier as NSObjectProtocol
     }
     
     func isEqual(toDiffableObject object: ListDiffable?) -> Bool {
-        guard let object = object as? LoadingCellViewModel
+        guard let object = object as? TextBasedCellViewModelType
             else { return false }
         
         return identifier == object.identifier
     }
     
 }
+